Abilities now Function on a Timer
The Zombie Fist and Ectoblast run on a timer, allowing you to use them for a set number of time rather than a set number of uses. The bonemerang still has one use though, and it does not run on a timer. You can also catch it by simply running into it now, giving you another use of it.